added favorites screen and fixed bug in parking lots (#9)
* added parking lots nearby and search options screens

* added results screen and sign up with google

* added favorites screen and fixed bug in parking lots nearby

* Revert "added favorites screen and fixed bug in parking lots nearby"

This reverts commit 473695d843f8fd7148467de9c887e5731248af8f.

* Revert "Revert "added favorites screen and fixed bug in parking lots nearby""

This reverts commit bdaad31f182df32ec2c02726d5e78103d75d52d3.

+++ lib/model/parking_lot.dart
@@ -6,7 +6,7 @@ class ParkingLot {
   final String opening_hours;
   final String fare;
   final int? is_accessible;
-  final int? availability;
+  final double? availability;
   final int? num_parking_spots;
   final String paying_method;
   final String? updated_time;
